Implement state machine for game logic
IState interface is used to create a class that represents
the game being in a certain state.
States will usually control a specific behavior such as moving a
player across the board after they roll.
States themselves control which state is active through the return
value of the update method.